# contracts/tester/v1/openapi.yaml
openapi: 3.0.3
info:
  title: TesterService API
  version: 0.0.1
paths:
  /problems:
    get:
      operationId: ListProblems
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            format: int32
          example: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
          example: 10
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProblemsResponse'
    post:
      operationId: CreateProblem
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProblemResponse'
  /problems/{id}:
    get:
      operationId: GetProblem
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProblemResponse'
    delete:
      operationId: DeleteProblem
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          example: 3
      responses:
        "200":
          description: OK
          content: { }
    patch:
      operationId: UpdateProblem
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          example: 3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProblemRequest'
      responses:
        "200":
          description: OK
          content: { }
    post:
      operationId: UploadProblem
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          example: 3
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadProblemRequest'
      responses:
        "200":
          description: OK
          content: { }
...
  /users:
    post:
      operationId: CreateUser
      security:
        - bearerAuth: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
    get:
      operationId: ListUsers
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            format: int32
          example: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
          example: 20
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'
  /users/{id}:
    get:
      operationId: GetUser
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          example: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserResponse'
    delete:
      operationId: DeleteUser
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          example: 3
      responses:
        "200":
          description: OK
          content: { }
    patch:
      operationId: UpdateUser
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          example: 3
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        "200":
          description: OK
          content: { }
components:
  schemas:
    Pagination:
      type: object
      required:
        - page
        - total
      properties:
        page:
          type: integer
          format: int32
          example: 1
        total:
          type: integer
          format: int32
          example: 10
    Problem:
      type: object
      required:
        - id
        - title
        - time_limit
        - memory_limit

        - legend
        - input_format
        - output_format
        - notes
        - scoring

        - legend_html
        - input_format_html
        - output_format_html
        - notes_html
        - scoring_html

        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int32
          example: 1
        title:
          type: string
          example: "Test problem"
        time_limit:
          type: integer
          format: int32
          example: 1000
        memory_limit:
          type: integer
          format: int32
          example: 256

        legend:
          type: string
        input_format:
          type: string
        output_format:
          type: string
        notes:
          type: string
        scoring:
          type: string

        legend_html:
          type: string
        input_format_html:
          type: string
        output_format_html:
          type: string
        notes_html:
          type: string
        scoring_html:
          type: string

        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProblemsListItem:
      type: object
      required:
        - id
        - title
        - memory_limit
        - time_limit
        - created_at
        - updated_at
        - solved_count
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        memory_limit:
          type: integer
          format: int32
        time_limit:
          type: integer
          format: int32
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        solved_count:
          type: integer
          format: int32
    ListProblemsResponse:
      type: object
      required:
        - problems
        - pagination
      properties:
        problems:
          type: array
          items:
            $ref: '#/components/schemas/ProblemsListItem'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateProblemResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int32
          example: 1
    GetProblemResponse:
      type: object
      required:
        - problem
      properties:
        problem:
          $ref: '#/components/schemas/Problem'
...
    UpdateProblemRequest:
      type: object
      properties:
        title:
          type: string
          example: "Test problem"
        legend:
          type: string
        input_format:
          type: string
        output_format:
          type: string
        notes:
          type: string
        scoring:
          type: string
        memory_limit:
          type: integer
          format: int32
          example: 256
        time_limit:
          type: integer
          format: int32
          example: 1000
    UploadProblemRequest:
      type: object
      required:
        - archive
      properties:
        archive:
          type: string
          format: binary
...
    CreateUserRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: user123
        password:
          type: string
          example: password123
    CreateUserResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int32
          example: 13
    GetUserResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/User'
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
          example: user123
        role:
          type: integer
          format: int32
          example: 1
    User:
      type: object
      required:
        - id
        - username
        - createdAt
        - modifiedAt
        - role
      properties:
        id:
          type: integer
          format: int32
          example: 13
        username:
          type: string
          example: user123
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
        role:
          type: integer
          format: int32
          example: 1
...
    ListUsersResponse:
      type: object
      required:
        - users
        - pagination
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT